requirements.txt
asgiref==3.6.0
coverage==7.2.7
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
flake8==6.0.0
flake8-html==0.4.3
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
